+++ resolwe_bio/tests/processes/test_amplicon.py
@@ -28,7 +28,17 @@ class AmpliconProcessorTestCase(BioProcessTestCase):
     def test_amplicon_table(self):
         bam = self.run_process('upload-bam', {'src': join('large', '56GSID_10k_mate1_RG.bam')})
         master_file = self.prepare_amplicon_master_file()
-        coverage = self.run_process('coveragebed', {'alignment': bam.id, 'master_file': master_file.id})
+        template = self.run_process('upload-file', {'src': 'report_html_template.html'})
+        bokeh_css = self.run_process('upload-file', {'src': 'bokeh-0.12.9.min.css'})
+        bokeh_js = self.run_process('upload-file', {'src': 'bokeh-0.12.9.min.js'})
+
+        coverage = self.run_process('coveragebed', {
+            'alignment': bam.id,
+            'master_file': master_file.id,
+            'template_html': template.id,
+            'bokeh_css': bokeh_css.id,
+            'bokeh_js': bokeh_js.id,
+        })
 
         inputs = {
             'annotation': '56GSID.lf.finalvars.txt',