+++ resolwe_bio/tests/processes/test_coverage.py
@@ -58,9 +58,22 @@ class CoverageProcessorTestCase(BioProcessTestCase):
 
     @skipUnlessLargeFiles('56GSID_10k_mate1_RG.bam')
     def test_amplicon_coverage(self):
+        template = self.run_process('upload-file', {'src': 'report_html_template.html'})
+        bokeh_css = self.run_process('upload-file', {'src': 'bokeh-0.12.9.min.css'})
+        bokeh_js = self.run_process('upload-file', {'src': 'bokeh-0.12.9.min.js'})
+
         bam = self.run_process('upload-bam', {'src': join('large', '56GSID_10k_mate1_RG.bam')})
         master_file = self.prepare_amplicon_master_file()
 
-        coverage = self.run_process('coveragebed', {'alignment': bam.id, 'master_file': master_file.id})
+        coverage = self.run_process('coveragebed', {
+            'alignment': bam.id,
+            'master_file': master_file.id,
+            'template_html': template.id,
+            'bokeh_css': bokeh_css.id,
+            'bokeh_js': bokeh_js.id,
+        })
         self.assertFile(coverage, 'cov_metrics', '56GSID_10k_covMetrics.txt')
         self.assertFile(coverage, 'mean_cov', '56GSID_10k_ampmeancov.covd')
+        self.assertFileExists(coverage, 'amplicon_cov')
+        self.assertFileExists(coverage, 'covplot')
+        self.assertFileExists(coverage, 'covplot_html')